+++ CHANGELOG.md
@@ -13,6 +13,8 @@ _Unreleased_
   `3.10.11`.  This can be accomplished by invoking `pin` with the new
   `--relaxed` flag.  #255
 
+- Default build-system and license can be specified in global config.  #244
+
 - Fixed an issue where the `init` command would not let you create
   `flit` based projects.  #254